requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements.txt --strip-extras requirements.in
#
docopt-ng==0.9.0
    # via staticjinja
jinja2==3.1.2
    # via staticjinja
libsass==0.22.0
    # via -r requirements.in
markupsafe==2.1.3
    # via jinja2
pyyaml==6.0
    # via -r requirements.in
staticjinja==5.0.0
    # via -r requirements.in
watchdog==3.0.0
    # via staticjinja
